requirements.txt
dateutils>=0.6.12
fastapi>=0.101.0
fastdup>=1.34
fonttools>=4.42.0
frozenlist>=1.4.0
fsspec>=2023.6.0
huggingface-hub>=0.16.4
hydra-core>=1.3.2
idna>=3.4
imageio>=2.31.1
inquirer>=3.1.3
ipykernel>=6.25.1
ipython>=8.14.0
ipywidgets>=8.1.0
iterative-telemetry>=0.0.8
itsdangerous>=2.1.2
Jinja2>=3.1.2
joblib>=1.3.1
jupyterlab-widgets>=3.0.8
kiwisolver>=1.4.4
kombu>=5.3.1
lazy_loader>=0.3
lightning>=2.0.6
lightning-cloud>=0.5.37
lightning-utilities>=0.9.0
lit>=16.0.6
Markdown>=3.4.4
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
matplotlib>=3.7.2
matplotlib-inline>=0.1.6
mdurl>=0.1.2
mpmath>=1.3.0
multidict>=6.0.4
nest-asyncio>=1.5.7
networkx>=3.1
numpy>=1.23.0
nvidia-cublas-cu11>=11.10.3.66
nvidia-cuda-cupti-cu11>=11.7.101
nvidia-cuda-nvrtc-cu11>=11.7.99
nvidia-cuda-runtime-cu11>=11.7.99
nvidia-cudnn-cu11>=8.5.0.96
nvidia-cufft-cu11>=10.9.0.58
nvidia-curand-cu11>=10.2.10.91
nvidia-cusolver-cu11>=11.4.0.1
nvidia-cusparse-cu11>=11.7.4.91
nvidia-nccl-cu11>=2.14.3
nvidia-nvtx-cu11>=11.7.91
oauthlib>=3.2.2
omegaconf>=2.3.0
opencv-python>=4.8.0.74
opencv-python-headless>=4.8.0.74
ordered-set>=4.1.0
orjson>=3.9.4
packaging>=23.1
pandas>=2.0.3
parso>=0.8.3
Pillow>=10.0.0
pillow-heif>=0.12.0
prompt-toolkit>=3.0.39
protobuf>=4.24.0
psutil>=5.9.5
pydot>=1.4.2
Pygments>=2.16.1
pygtrie>=2.5.0
pyparsing>=3.0.9
python-dateutil>=2.8.2
python-editor>=1.0.4
python-multipart>=0.0.6
pytorch-lightning>=2.0.6
pytorch-ood>=0.1.5
pytz>=2023.3
PyWavelets>=1.4.1
PyYAML>=6.0.1
pyzmq>=25.1.1
readchar>=4.0.5
regex>=2023.8.8
requests>=2.28.1
requests-oauthlib>=1.3.1
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.1
scmrepo>=1.1.0
seaborn>=0.12.2
sentry-sdk>=1.29.2
shortuuid>=1.0.11
shtab>=1.6.4
tensorboard>=2.14.0
tensorboard-data-server>=0.7.1
tokenizers>=0.13.3
tomlkit>=0.12.1
torch>=2.0.1+cu118
torchaudio>=2.0.2+cu118
torchmetrics>=1.0.3
torchvision>=0.15.2+cu118
tornado>=6.3.3
tqdm>=4.65.2
traitlets>=5.9.0
transformers>=4.31.0
triton>=2.0.0
typing_extensions>=4.7.1
tzdata>=2023.3
urllib3>=1.26.16
Werkzeug>=2.3.6
widgetsnbextension>=4.0.8